refactor: Set modern portfolio as main page
This commit restructures the application's main entry point.

- The content previously in `pages/modern-portfolio/index.tsx` is now in `pages/index.tsx`, making the modern portfolio design the default page for the application. This new main page includes the "Web Applications for Local Businesses" section.
- The original content of `pages/index.tsx` (using `PortfolioDisplay`) has been moved to `pages/classic.tsx`.
- Due to persistent build errors when integrating the new "Web Applications for Local Businesses" section into `app/component/portfolio/portfolioDisplay.tsx`, I reverted this component to its original state. As a result, the `pages/classic.tsx` page does NOT currently include the new business section. This can be addressed in a future update if needed.
- I have removed the `pages/modern-portfolio/` directory.

The application now builds successfully with these changes.
